fix build issues

+++ src/engraving/utests/playbackeventsrendering_tests.cpp
@@ -360,9 +360,9 @@ TEST_F(PlaybackEventsRendererTests, SingleNote_Turn_Regular)
     int expectedSubNotesCount = 4;
     duration_t expectedDuration = QUARTER_NOTE_DURATION / expectedSubNotesCount;
     pitch_level_t nominalPitchLevel = pitchLevel(PitchClass::F, 4);
-    std::vector<pitch_level_t> expectedPitches = { nominalPitchLevel + PITCH_LEVEL_STEP,
+    std::vector<pitch_level_t> expectedPitches = { static_cast<short>(nominalPitchLevel + PITCH_LEVEL_STEP),
                                                    nominalPitchLevel,
-                                                   nominalPitchLevel - PITCH_LEVEL_STEP,
+                                                   static_cast<short>(nominalPitchLevel - PITCH_LEVEL_STEP),
                                                    nominalPitchLevel };
 
     // [GIVEN] Fulfill articulations profile with dummy patterns
@@ -414,9 +414,9 @@ TEST_F(PlaybackEventsRendererTests, SingleNote_Turn_Inverted)
     int expectedSubNotesCount = 4;
     duration_t expectedDuration = QUARTER_NOTE_DURATION / expectedSubNotesCount;
     pitch_level_t nominalPitchLevel = pitchLevel(PitchClass::F, 4);
-    std::vector<pitch_level_t> expectedPitches = { nominalPitchLevel - PITCH_LEVEL_STEP,
+    std::vector<pitch_level_t> expectedPitches = { static_cast<short>(nominalPitchLevel - PITCH_LEVEL_STEP),
                                                    nominalPitchLevel,
-                                                   nominalPitchLevel + PITCH_LEVEL_STEP,
+                                                   static_cast<short>(nominalPitchLevel + PITCH_LEVEL_STEP),
                                                    nominalPitchLevel };
 
     // [GIVEN] Fulfill articulations profile with dummy patterns
@@ -470,9 +470,9 @@ TEST_F(PlaybackEventsRendererTests, SingleNote_Turn_Inverted_Slash_Variation)
     int expectedSubNotesCount = 4;
     duration_t expectedDuration = QUARTER_NOTE_DURATION / expectedSubNotesCount;
     pitch_level_t nominalPitchLevel = pitchLevel(PitchClass::F, 4);
-    std::vector<pitch_level_t> expectedPitches = { nominalPitchLevel - PITCH_LEVEL_STEP,
+    std::vector<pitch_level_t> expectedPitches = { static_cast<short>(nominalPitchLevel - PITCH_LEVEL_STEP),
                                                    nominalPitchLevel,
-                                                   nominalPitchLevel + PITCH_LEVEL_STEP,
+                                                   static_cast<short>(nominalPitchLevel + PITCH_LEVEL_STEP),
                                                    nominalPitchLevel };
 
     // [GIVEN] Fulfill articulations profile with dummy patterns
@@ -539,7 +539,7 @@ TEST_F(PlaybackEventsRendererTests, SingleNote_Upper_Mordent)
 
     std::vector<pitch_level_t> expectedPitches = {
         nominalPitchLevel,
-        nominalPitchLevel + PITCH_LEVEL_STEP,
+        static_cast<short>(nominalPitchLevel + PITCH_LEVEL_STEP),
         nominalPitchLevel
     };
 
@@ -607,7 +607,7 @@ TEST_F(PlaybackEventsRendererTests, SingleNote_Lower_Mordent)
 
     std::vector<pitch_level_t> expectedPitches = {
         nominalPitchLevel,
-        nominalPitchLevel - PITCH_LEVEL_STEP,
+        static_cast<short>(nominalPitchLevel - PITCH_LEVEL_STEP),
         nominalPitchLevel
     };
